fix: feedback revisions, r3
